Adding Drawer. Converting Prayer to Fragment
