+++ README.md
@@ -1,0 +1,18 @@
+# PrayerApp
+An attempt at a prayer app, initially aimed at the Church In Wales
+
+This app is an attempt to bring the Church In Wales Daily Offices to a mobile device. 
+
+It can currently:
+
+1. Offer a Morning Prayer (including a cycle of readings) in both English and Welsh.
+
+#Downloading
+
+To download the current version of the app you can get it form here : https://github.com/RevBettinson/PrayerApp/releases/download/ALPHAv0.0002/ChurchInWalesPrayerApp.apk
+
+Your device may warn you about installation from "uknown sources", click "Okay". Make sure that you only do it for this app, and not for everything, if that's an option. 
+Your device may warn you about "Unknown" or "Untrusted" publisher. Click "Okay" or "Install Anway" (depending on the phone). 
+
+You should find the app called "Prayer" is now installed. It has a little Android face as the icon.
+
