+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion2.txt
@@ -1,0 +1,12 @@
+<h2>The Conclusion</h2>
+<br>
+The Lord bless us and protect us from evil
+and keep us in eternal life. <strong>Amen</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
