+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion3.txt
@@ -1,0 +1,15 @@
+<h2>The Conclusion</h2>
+<br>
+Now to him who is able through the power which is at 
+work among us to do immeasurably more than we can ask 
+or conceive, to him be glory in the church and in Christ 
+Jesus from generation to generation for ever. <strong>Amen.</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
+
