+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion4.txt
@@ -1,0 +1,12 @@
+<h2>The Conclusion</h2>
+<br>
+To the King eternal, immortal, invisible, the only God, be
+honour and glory for ever. <strong>Amen</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
