+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion5.txt
@@ -1,0 +1,15 @@
+<h2>The Conclusion</h2>
+<br>
+Now to him who can keep us from falling and set us in
+the presence of his glory, jubilant and above reproach, to
+the only God our Saviour, be glory and majesty, power and
+authority, through Jesus Christ our Lord, before all time,
+now and for ever. <strong>Amen</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
