+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion6.txt
@@ -1,0 +1,14 @@
+<h2>The Conclusion</h2>
+<br>
+To him who loves us and has set us free from our sins
+with his blood, who has made us a royal house, to serve
+as the priests of his God and Father: to him be glory and
+dominion for ever. <strong>Amen</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
