+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion7.txt
@@ -1,0 +1,13 @@
+<h2>The Conclusion</h2>
+<br>
+To the holy and glorious Trinity, Father, Son, and Holy
+Spirit, be blessing and honour, thanksgiving and praise,
+now and for ever. <strong>Amen.</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
