+++ app/src/main/Prayer/MorningPrayer/Conclusions/EN_Conclusion8.txt
@@ -1,0 +1,19 @@
+<h2>The Conclusion</h2>
+<br>
+To God the Father, who first loved us,
+and made us accepted in the Beloved;
+to God the Son, who loved us,
+and washed us from our sins by his own blood;
+to God the Holy Spirit,
+who sheds the love of God abroad in our hearts;
+to the one true God
+be all love and all glory,
+for time and for eternity. <strong>Amen.</strong>
+<br>
+<br>
+The Lord be with you.<br>
+<strong>And also with you.</strong><br>
+<br>
+Let us bless the Lord.<br>
+<strong>Thanks be to God.</strong>
+<br>
