+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal1.txt
@@ -1,0 +1,3 @@
+The peace of God which is beyond all understanding
+ guard your hearts in the knowledge and love of God and
+ of his Son Jesus Christ …
