+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal2.txt
@@ -1,0 +1,3 @@
+The Lord bless you and keep watch over you; the Lord’s
+ face shine on you and be gracious to you; the Lord look
+ lovingly on you and give you peace …
