+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal3.txt
@@ -1,0 +1,2 @@
+The God of all grace who called you to his eternal glory in
+ Christ, restore, establish and strengthen you in the Faith …
