+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal4.txt
@@ -1,0 +1,3 @@
+God, who from the death of sin raised you to new life in
+ Christ, keep you from falling and set you in the presence
+ of his glory with great joy …
