+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal5.txt
@@ -1,0 +1,3 @@
+Christ, who nourishes you with himself, the living bread and
+ the true vine, make you one in praise and service, and
+ raise you up at the last day …
