+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal6.txt
@@ -1,0 +1,3 @@
+The God of hope fill you with all joy and peace in
+ believing, that by the power of the Holy Spirit you may be
+ filled with hope …
