+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal7.txt
@@ -1,0 +1,3 @@
+The love of the Lord Jesus draw you to himself; the power
+ of the Lord Jesus strengthen you in his service; the joy of
+ the Lord Jesus fill your hearts …
