+++ app/src/main/Prayer/MorningPrayer/Dismissals/EN_Dismissal8.txt
@@ -1,0 +1,3 @@
+Keep your eyes fixed on Jesus who was wounded for our sins,
+ that you may bear in your life the love and joy and peace
+ which are the marks of Jesus in his disciples …
