+++ app/src/main/res/values/themes.xml
@@ -13,4 +13,7 @@
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
     </style>
+
+
+    <style name="Theme.Prayer.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 </resources>
